Create new-rest-api-on-books.go
The web App has four Endpoints: an Endpoint for returning all books, an Endpoint for returning a single book, an Endpoint for creating a new book, an Endpoint for deleting a particular book by id. The Gorilla mux package, has been used for creating routes. Each Endpoint can be tested with either; Postman, or Http Client, or by the command: "curl" in a command line.