Create reviewed-rest-api-on-books.go
The Web App is a Resful API of which data is encoded in the JSON format. The App has five endpoints: an Endpoint for creating a new book. An Endpoint for returning all the books. An Endpoint for returning a single book. An Endpoint for deleting a book. An endpoint for updating. The gorilla mux  package has been applied for Routing over the Endpoints. The App Endpoints can be tested with Postman, or Http Client, or with the command: "curl", on the Terminal.